SiteCreate.php - Allow separating the download and install steps
The site:create command currently performs a download (from tarball) and
installation (setting up DB and config files).  That is handy for
quick-starting, but when piecing-together different dev workflows, it's
helpful to have these as separate steps.

The change aims to maintain the current "site:create" command while
still allowing download/install to be split apart.